+++ src/quantum/azext_quantum/operations/job.py
@@ -7,7 +7,8 @@
 
 import logging
 
-from knack.util import CLIError
+from azure.cli.core.azclierror import (FileOperationError, AzureInternalError,
+                                       InvalidArgumentValueError, AzureResponseError)
 
 from .._client_factory import cf_jobs, _get_data_credentials, base_url
 from .workspace import WorkspaceInfo
@@ -43,11 +44,11 @@ def _check_dotnet_available():
     try:
         import subprocess
         result = subprocess.run(args, stdout=subprocess.PIPE, check=False)
-    except FileNotFoundError:
-        raise CLIError(f"Could not find 'dotnet' on the system.")
+    except FileNotFoundError as e:
+        raise FileOperationError("Could not find 'dotnet' on the system.") from e
 
     if result.returncode != 0:
-        raise CLIError(f"Failed to run 'dotnet'. (Error {result.returncode})")
+        raise FileOperationError(f"Failed to run 'dotnet'. (Error {result.returncode})")
 
 
 def build(cmd, target_id=None, project=None):
@@ -78,7 +79,7 @@ def build(cmd, target_id=None, project=None):
     # If we got here, we might have encountered an error during compilation, so propagate standard output to the user.
     logger.error(f"Compilation stage failed with error code {result.returncode}")
     print(result.stdout.decode('ascii'))
-    raise CLIError("Failed to compile program.")
+    raise AzureInternalError("Failed to compile program.")
 
 
 def _generate_submit_args(program_args, ws, target, token, project, job_name, shots, storage, job_params):
@@ -125,6 +126,9 @@ def _generate_submit_args(program_args, ws, target, token, project, job_name, sh
 
     args.append("--location")
     args.append(ws.location)
+
+    args.append("--user-agent")
+    args.append("CLI")
 
     if job_params:
         args.append("--job-params")
@@ -190,7 +194,7 @@ def submit(cmd, program_args, resource_group_name=None, workspace_name=None, loc
     # The program compiled succesfully, but executing the stand-alone .exe failed to run.
     logger.error(f"Submission of job failed with error code {result.returncode}")
     print(result.stdout.decode('ascii'))
-    raise CLIError("Failed to submit job.")
+    raise AzureInternalError("Failed to submit job.")
 
 
 def _parse_blob_url(url):
@@ -202,8 +206,8 @@ def _parse_blob_url(url):
         container = o.path.split('/')[-2]
         blob = o.path.split('/')[-1]
         sas_token = o.query
-    except IndexError:
-        raise CLIError(f"Failed to parse malformed blob URL: {url}")
+    except IndexError as e:
+        raise InvalidArgumentValueError(f"Failed to parse malformed blob URL: {url}") from e
 
     return {
         "account_name": account_name,
@@ -253,7 +257,7 @@ def output(cmd, job_id, resource_group_name=None, workspace_name=None, location=
                 while result_start_line >= 0 and not lines[result_start_line].startswith('"'):
                     result_start_line -= 1
             if result_start_line < 0:
-                raise CLIError("Job output is malformed, mismatched quote characters.")
+                raise AzureResponseError("Job output is malformed, mismatched quote characters.")
 
             # Print the job output and then the result of the operation as a histogram.
             # If the result is a string, trim the quotation marks.